Build Page 생성
1. Databinding 적용
2. BuildViewmodel생성
3. RecycleView에 Data 삽입
