+++ app/build.gradle
@@ -4,6 +4,7 @@ plugins {
     id 'kotlin-kapt'
     id 'dagger.hilt.android.plugin'
     id 'androidx.navigation.safeargs.kotlin'
+    id 'kotlin-parcelize'
 }
 
 android {
@@ -126,4 +127,7 @@ dependencies {
 
     //sandwich
     implementation "com.github.skydoves:sandwich:1.2.1"
+
+    //whatif
+    implementation "com.github.skydoves:whatif:1.1.0"
 }
