+++ app/src/main/java/com/example/firstapp/model/Item.kt
@@ -4,6 +4,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.Ignore
 import androidx.room.PrimaryKey
+import com.example.firstapp.model.Champion.Image
 
 data class ItemJson(
     val type: String,
